Creating the game world and the hat Script
The Game world has been created from the sprites given (Downloaded from
Unity Store), and set up as the Artwork in the Main scene. Right after
positioning the sprites, the hat object was given a rigidbody 2D, and
is kinematic in order to stop the hat from being effected by gravity.