Using the Camera script and fixed update instead of update
The camera variable cam is set to public that means it’s accessible
from all scripts in the game. Also the fixed update is now called once
per physics timestep rather than once per frame which can be used in
the physics settings.

The ScreenToWorldPoint will work on both mouse and smartphones or on
most smartphones as the code works directly on the software itself.