Hat positioning on screen
Hat is set to a boundary against the far end of the screen, therefore
we assign a renderer to prove the size of the hat according to the
movement of the mouse. This will make the hat hit the boundary and
stop, which will allow the user to see the full hat displayed from the
edge of the object not from the centre position assigned in the
beginning of the scene input.