Adding assets
Bowling balls have been added to the scene , adding a Physics RigidBody
2D also but without using the kinematic option. This allows the bowling
ball fall together with the gravity therefore if the bowling ball is
not captured by the hat, no points are added.

The edge collider 2D is added to the hat since we need a “location” to
where the hat is captured. Edit collider gives the chance to edit the
“mesh” that needs to be collided to.

Also the hat sprite is divided into three different parts in photoshop,
the front of the hat and the top rear side of the hat separate. This
will allow the bowling ball to fall directly “into” the hat while
editing the order in layers, placing the top rear a layer after the
bowling ball and the rest of the hat before the bowling ball, which
means it will cover the bowling ball whilst entering the hat.