Scripts, Start & Restart
Right after the bowling ball is being able to hit the collider
(HatBackSprite), the next thing was to get the bombs up and running in
a random position and also during a random time. Once the bomb ball is
dropped, the the DestroyOnContact script is active and calls for the
Explosion from the Materials set from the prefab.

Points & Timer
Points are controlled in the game controller that is specifically
reading the collider from the hat whilst picking the next object
falling. If the bowling ball is falling hitting the hat’s collider.
Once the object hat reaches the collider, a point is added to the
score, if the player fails to pick the bowling ball no points are give.
On the other hand when the bombs hit the collider 2 points are deducted
from the score.

IENumerator has been used to respawn the bowling balls and also the
bombs whilst having the ScreenToWorldPoint that determined the area
space a bowling ball and the bomb will fall, measuring the whole screen
for the to be able to collect the bowling ball. This is the same thing
when the hat was set to be displayed at all time given edge (Screen
limit). The timer has been set by me at 10 seconds, set in the
GameController ‘TimeLeft’. (In Future i would add a game option for the
user to choose between different timers.)

Start, Splash and Restart

The start button is is assigned to the HT_StartGame script, and calls
for an ‘onmousedown’ action by the user which will load the
GameController and starts the game scene. Also the timer once the 10
seconds are up, it will display a Game Over label, and 2 seconds after
displays the restart button. The reset button will recall the scene
from the start, resetting the score and timer.